[알고리즘] 비밀지도 _프로그래머스  완성 2,6번 테스트 런타임 에러 해결 Integer -> Long

+++ src/Study221107_221111/Study221111/SecretMap.java
@@ -5,7 +5,7 @@ public class SecretMap {
         String[] answer = new String[n];
 
         for (int i = 0; i < n; i++) {
-            answer[i] = String.format("%0"+n+"d", Integer.parseInt(Integer.toBinaryString(arr1[i] | arr2[i])));
+            answer[i] = String.format("%0"+n+"d", Long.parseLong(Long.toBinaryString(arr1[i] | arr2[i])));
             answer[i] = answer[i].replace("1", "#");
             answer[i] = answer[i].replace("0", " ");
             System.out.println(answer[i]);
